remove CWJ for now from common
